consolidate test fixtures into conftest files
